requirements.txt
pymata4
transitions
paho-mqtt
